se eliminó el componente Cards
